+++ taskflowbackend/tests/base.py
@@ -57,9 +57,7 @@ PROJECT_TYPE_PROJECT = SODAR_CONSTANTS['PROJECT_TYPE_PROJECT']
 APP_SETTING_SCOPE_PROJECT = SODAR_CONSTANTS['APP_SETTING_SCOPE_PROJECT']
 
 # Local constants
-IRODS_ACCESS_READ = 'read object'
 IRODS_ACCESS_OWN = 'own'
-IRODS_ACCESS_WRITE = 'modify object'
 IRODS_ACCESS_NULL = 'null'
 IRODS_GROUP_PUBLIC = 'public'
 TICKET_STR = 'ei8iomuDoazeiD2z'
@@ -268,6 +266,11 @@ class TaskflowTestMixin(ProjectMixin, RoleMixin, RoleAssignmentMixin):
         self.owner_as_cat = self.make_assignment(
             self.category, self.user_owner_cat, self.role_owner
         )
+        # Set iRODS 4.2/4.3 compatible ACL params
+        new_ver = self.irods_backend.is_irods_version(self.irods, '4.3')
+        acl_dl = '_' if new_ver else ' '
+        self.irods_access_read = 'read{}object'.format(acl_dl)
+        self.irods_access_write = 'modify{}object'.format(acl_dl)
 
     def tearDown(self):
         self.clear_irods_test_data()