+++ taskflowbackend/tests/test_tasks.py
@@ -44,12 +44,11 @@ TEST_VAL = 'test_val'
 TEST_UNITS = 'test_units'
 TEST_USER_GROUP = USER_PREFIX + 'group2'
 
-# NOTE: Yes, we really need this for the python irods client
-TEST_ACCESS_READ_IN = 'read'
-TEST_ACCESS_READ_OUT = 'read object'
-TEST_ACCESS_WRITE_IN = 'write'
-TEST_ACCESS_WRITE_OUT = 'modify object'
-TEST_ACCESS_NULL = 'null'
+# iRODS access control values
+# NOTE: input values set in base class for iRODS 4.2/4.3 support
+IRODS_ACCESS_READ_IN = 'read'
+IRODS_ACCESS_WRITE_IN = 'write'
+IRODS_ACCESS_NULL = 'null'
 
 BATCH_SRC_NAME = 'batch_src'
 BATCH_DEST_NAME = 'batch_dest'
@@ -776,9 +775,10 @@ class TestSetAccessTask(IRODSTaskTestBase):
             cls=SetAccessTask,
             name='Set access',
             inject={
-                'access_name': TEST_ACCESS_READ_IN,
+                'access_name': IRODS_ACCESS_READ_IN,
                 'path': self.test_coll_path,
                 'user_name': DEFAULT_USER_GROUP,
+                'irods_backend': self.irods_backend,
             },
         )
         user_access = self._get_user_access(
@@ -792,7 +792,7 @@ class TestSetAccessTask(IRODSTaskTestBase):
             target=self._get_test_coll(), user_name=DEFAULT_USER_GROUP
         )
         self.assertIsInstance(user_access, iRODSAccess)
-        self.assertEqual(user_access.access_name, TEST_ACCESS_READ_OUT)
+        self.assertEqual(user_access.access_name, self.irods_access_read)
 
     def test_execute_write(self):
         """Test access setting for write/modify"""
@@ -800,9 +800,10 @@ class TestSetAccessTask(IRODSTaskTestBase):
             cls=SetAccessTask,
             name='Set access',
             inject={
-                'access_name': TEST_ACCESS_WRITE_IN,
+                'access_name': IRODS_ACCESS_WRITE_IN,
                 'path': self.test_coll_path,
                 'user_name': DEFAULT_USER_GROUP,
+                'irods_backend': self.irods_backend,
             },
         )
         user_access = self._get_user_access(
@@ -817,7 +818,7 @@ class TestSetAccessTask(IRODSTaskTestBase):
             target=self._get_test_coll(), user_name=DEFAULT_USER_GROUP
         )
         self.assertIsInstance(user_access, iRODSAccess)
-        self.assertEqual(user_access.access_name, TEST_ACCESS_WRITE_OUT)
+        self.assertEqual(user_access.access_name, self.irods_access_write)
 
     def test_execute_twice(self):
         """Test access setting twice"""
@@ -825,9 +826,10 @@ class TestSetAccessTask(IRODSTaskTestBase):
             cls=SetAccessTask,
             name='Set access',
             inject={
-                'access_name': TEST_ACCESS_READ_IN,
+                'access_name': IRODS_ACCESS_READ_IN,
                 'path': self.test_coll_path,
                 'user_name': DEFAULT_USER_GROUP,
+                'irods_backend': self.irods_backend,
             },
         )
         result = self._run_flow()
@@ -838,9 +840,10 @@ class TestSetAccessTask(IRODSTaskTestBase):
             cls=SetAccessTask,
             name='Set access',
             inject={
-                'access_name': TEST_ACCESS_READ_IN,
+                'access_name': IRODS_ACCESS_READ_IN,
                 'path': self.test_coll_path,
                 'user_name': DEFAULT_USER_GROUP,
+                'irods_backend': self.irods_backend,
             },
         )
         result = self._run_flow()
@@ -850,7 +853,7 @@ class TestSetAccessTask(IRODSTaskTestBase):
             target=self._get_test_coll(), user_name=DEFAULT_USER_GROUP
         )
         self.assertIsInstance(user_access, iRODSAccess)
-        self.assertEqual(user_access.access_name, TEST_ACCESS_READ_OUT)
+        self.assertEqual(user_access.access_name, self.irods_access_read)
 
     def test_revert_created(self):
         """Test access setting"""
@@ -858,9 +861,10 @@ class TestSetAccessTask(IRODSTaskTestBase):
             cls=SetAccessTask,
             name='Set access',
             inject={
-                'access_name': TEST_ACCESS_READ_IN,
+                'access_name': IRODS_ACCESS_READ_IN,
                 'path': self.test_coll_path,
                 'user_name': DEFAULT_USER_GROUP,
+                'irods_backend': self.irods_backend,
             },
             force_fail=True,
         )  # FAIL
@@ -879,9 +883,10 @@ class TestSetAccessTask(IRODSTaskTestBase):
             cls=SetAccessTask,
             name='Set access',
             inject={
-                'access_name': TEST_ACCESS_READ_IN,
+                'access_name': IRODS_ACCESS_READ_IN,
                 'path': self.test_coll_path,
                 'user_name': DEFAULT_USER_GROUP,
+                'irods_backend': self.irods_backend,
             },
         )
         self._run_flow()
@@ -891,9 +896,10 @@ class TestSetAccessTask(IRODSTaskTestBase):
             cls=SetAccessTask,
             name='Set access',
             inject={
-                'access_name': TEST_ACCESS_WRITE_IN,
+                'access_name': IRODS_ACCESS_WRITE_IN,
                 'path': self.test_coll_path,
                 'user_name': DEFAULT_USER_GROUP,
+                'irods_backend': self.irods_backend,
             },
             force_fail=True,
         )  # FAIL
@@ -904,7 +910,7 @@ class TestSetAccessTask(IRODSTaskTestBase):
             target=self._get_test_coll(), user_name=DEFAULT_USER_GROUP
         )
         self.assertIsInstance(user_access, iRODSAccess)
-        self.assertEqual(user_access.access_name, TEST_ACCESS_READ_OUT)
+        self.assertEqual(user_access.access_name, self.irods_access_read)
 
     def test_revert_not_modified(self):
         """Test access setting reverting without modification"""
@@ -912,9 +918,10 @@ class TestSetAccessTask(IRODSTaskTestBase):
             cls=SetAccessTask,
             name='Set access',
             inject={
-                'access_name': TEST_ACCESS_READ_IN,
+                'access_name': IRODS_ACCESS_READ_IN,
                 'path': self.test_coll_path,
                 'user_name': DEFAULT_USER_GROUP,
+                'irods_backend': self.irods_backend,
             },
         )
         result = self._run_flow()
@@ -925,9 +932,10 @@ class TestSetAccessTask(IRODSTaskTestBase):
             cls=SetAccessTask,
             name='Set access',
             inject={
-                'access_name': TEST_ACCESS_READ_IN,
+                'access_name': IRODS_ACCESS_READ_IN,
                 'path': self.test_coll_path,
                 'user_name': DEFAULT_USER_GROUP,
+                'irods_backend': self.irods_backend,
             },
             force_fail=True,
         )  # FAIL
@@ -938,7 +946,7 @@ class TestSetAccessTask(IRODSTaskTestBase):
             target=self._get_test_coll(), user_name=DEFAULT_USER_GROUP
         )
         self.assertIsInstance(user_access, iRODSAccess)
-        self.assertEqual(user_access.access_name, TEST_ACCESS_READ_OUT)
+        self.assertEqual(user_access.access_name, self.irods_access_read)
 
     def test_execute_no_recursion(self):
         """Test access setting for a collection with recursive=False"""
@@ -953,9 +961,10 @@ class TestSetAccessTask(IRODSTaskTestBase):
             cls=SetAccessTask,
             name='Set access',
             inject={
-                'access_name': TEST_ACCESS_READ_IN,
+                'access_name': IRODS_ACCESS_READ_IN,
                 'path': self.test_coll_path,
                 'user_name': TEST_USER,
+                'irods_backend': self.irods_backend,
                 'recursive': False,
             },
         )
@@ -976,7 +985,7 @@ class TestSetAccessTask(IRODSTaskTestBase):
             target=self._get_test_coll(), user_name=TEST_USER
         )
         self.assertIsInstance(user_access, iRODSAccess)
-        self.assertEqual(user_access.access_name, TEST_ACCESS_READ_OUT)
+        self.assertEqual(user_access.access_name, self.irods_access_read)
 
         user_access = self._get_user_access(
             target=sub_coll, user_name=TEST_USER
@@ -996,9 +1005,10 @@ class TestSetAccessTask(IRODSTaskTestBase):
             cls=SetAccessTask,
             name='Set access',
             inject={
-                'access_name': TEST_ACCESS_READ_IN,
+                'access_name': IRODS_ACCESS_READ_IN,
                 'path': self.test_coll_path,
                 'user_name': TEST_USER,
+                'irods_backend': self.irods_backend,
                 'recursive': False,
             },
             force_fail=True,
@@ -1041,7 +1051,7 @@ class TestIssueTicketTask(IRODSTaskTestBase):
             cls=IssueTicketTask,
             name='Issue ticket',
             inject={
-                'access_name': TEST_ACCESS_READ_IN,
+                'access_name': IRODS_ACCESS_READ_IN,
                 'path': self.test_coll_path,
                 'ticket_str': TICKET_STR,
                 'irods_backend': self.irods_backend,
@@ -1059,7 +1069,7 @@ class TestIssueTicketTask(IRODSTaskTestBase):
             cls=IssueTicketTask,
             name='Issue ticket',
             inject={
-                'access_name': TEST_ACCESS_READ_IN,
+                'access_name': IRODS_ACCESS_READ_IN,
                 'path': self.test_coll_path,
                 'ticket_str': TICKET_STR,
                 'irods_backend': self.irods_backend,
@@ -1076,7 +1086,7 @@ class TestIssueTicketTask(IRODSTaskTestBase):
             cls=IssueTicketTask,
             name='Issue ticket',
             inject={
-                'access_name': TEST_ACCESS_READ_IN,
+                'access_name': IRODS_ACCESS_READ_IN,
                 'path': self.test_coll_path,
                 'ticket_str': TICKET_STR,
                 'irods_backend': self.irods_backend,
@@ -1095,7 +1105,7 @@ class TestIssueTicketTask(IRODSTaskTestBase):
             cls=IssueTicketTask,
             name='Issue ticket',
             inject={
-                'access_name': TEST_ACCESS_READ_IN,
+                'access_name': IRODS_ACCESS_READ_IN,
                 'path': self.test_coll_path,
                 'ticket_str': TICKET_STR,
                 'irods_backend': self.irods_backend,
@@ -1109,7 +1119,7 @@ class TestIssueTicketTask(IRODSTaskTestBase):
     def test_revert_not_modified(self):
         """Test reverting a ticket issuing with no modification"""
         self.irods_backend.issue_ticket(
-            self.irods, TEST_ACCESS_READ_IN, self.test_coll_path, TICKET_STR
+            self.irods, IRODS_ACCESS_READ_IN, self.test_coll_path, TICKET_STR
         )
         self.assertIsInstance(
             self.irods_backend.get_ticket(self.irods, TICKET_STR), Ticket
@@ -1118,7 +1128,7 @@ class TestIssueTicketTask(IRODSTaskTestBase):
             cls=IssueTicketTask,
             name='Issue ticket',
             inject={
-                'access_name': TEST_ACCESS_READ_IN,
+                'access_name': IRODS_ACCESS_READ_IN,
                 'path': self.test_coll_path,
                 'ticket_str': TICKET_STR,
                 'irods_backend': self.irods_backend,
@@ -1138,7 +1148,7 @@ class TestDeleteTicketTask(IRODSTaskTestBase):
     def test_execute(self):
         """Test deleting a ticket"""
         self.irods_backend.issue_ticket(
-            self.irods, TEST_ACCESS_READ_IN, self.test_coll_path, TICKET_STR
+            self.irods, IRODS_ACCESS_READ_IN, self.test_coll_path, TICKET_STR
         )
         self.assertIsInstance(
             self.irods_backend.get_ticket(self.irods, TICKET_STR), Ticket
@@ -1147,7 +1157,7 @@ class TestDeleteTicketTask(IRODSTaskTestBase):
             cls=DeleteTicketTask,
             name='Delete ticket',
             inject={
-                'access_name': TEST_ACCESS_READ_IN,
+                'access_name': IRODS_ACCESS_READ_IN,
                 'path': self.test_coll_path,
                 'ticket_str': TICKET_STR,
                 'irods_backend': self.irods_backend,
@@ -1160,7 +1170,7 @@ class TestDeleteTicketTask(IRODSTaskTestBase):
     def test_execute_twice(self):
         """Test deleting a ticket twice"""
         self.irods_backend.issue_ticket(
-            self.irods, TEST_ACCESS_READ_IN, self.test_coll_path, TICKET_STR
+            self.irods, IRODS_ACCESS_READ_IN, self.test_coll_path, TICKET_STR
         )
         self.assertIsInstance(
             self.irods_backend.get_ticket(self.irods, TICKET_STR), Ticket
@@ -1169,7 +1179,7 @@ class TestDeleteTicketTask(IRODSTaskTestBase):
             cls=DeleteTicketTask,
             name='Delete ticket',
             inject={
-                'access_name': TEST_ACCESS_READ_IN,
+                'access_name': IRODS_ACCESS_READ_IN,
                 'path': self.test_coll_path,
                 'ticket_str': TICKET_STR,
                 'irods_backend': self.irods_backend,
@@ -1184,7 +1194,7 @@ class TestDeleteTicketTask(IRODSTaskTestBase):
             cls=DeleteTicketTask,
             name='Delete ticket',
             inject={
-                'access_name': TEST_ACCESS_READ_IN,
+                'access_name': IRODS_ACCESS_READ_IN,
                 'path': self.test_coll_path,
                 'ticket_str': TICKET_STR,
                 'irods_backend': self.irods_backend,
@@ -1197,7 +1207,7 @@ class TestDeleteTicketTask(IRODSTaskTestBase):
     def test_revert_modified(self):
         """Test reverting ticket deletion"""
         self.irods_backend.issue_ticket(
-            self.irods, TEST_ACCESS_READ_IN, self.test_coll_path, TICKET_STR
+            self.irods, IRODS_ACCESS_READ_IN, self.test_coll_path, TICKET_STR
         )
         self.assertIsInstance(
             self.irods_backend.get_ticket(self.irods, TICKET_STR), Ticket
@@ -1206,7 +1216,7 @@ class TestDeleteTicketTask(IRODSTaskTestBase):
             cls=DeleteTicketTask,
             name='Delete ticket',
             inject={
-                'access_name': TEST_ACCESS_READ_IN,
+                'access_name': IRODS_ACCESS_READ_IN,
                 'path': self.test_coll_path,
                 'ticket_str': TICKET_STR,
                 'irods_backend': self.irods_backend,
@@ -1226,7 +1236,7 @@ class TestDeleteTicketTask(IRODSTaskTestBase):
             cls=DeleteTicketTask,
             name='Delete ticket',
             inject={
-                'access_name': TEST_ACCESS_READ_IN,
+                'access_name': IRODS_ACCESS_READ_IN,
                 'path': self.test_coll_path,
                 'ticket_str': TICKET_STR,
                 'irods_backend': self.irods_backend,
@@ -1258,9 +1268,10 @@ class TestSetDataObjAccessTask(IRODSTaskTestBase):
             cls=SetAccessTask,
             name='Set access',
             inject={
-                'access_name': TEST_ACCESS_READ_IN,
+                'access_name': IRODS_ACCESS_READ_IN,
                 'path': self.obj_path,
                 'user_name': DEFAULT_USER_GROUP,
+                'irods_backend': self.irods_backend,
                 'obj_target': True,
             },
         )
@@ -1276,7 +1287,7 @@ class TestSetDataObjAccessTask(IRODSTaskTestBase):
             target=self._get_test_obj(), user_name=DEFAULT_USER_GROUP
         )
         self.assertIsInstance(user_access, iRODSAccess)
-        self.assertEqual(user_access.access_name, TEST_ACCESS_READ_OUT)
+        self.assertEqual(user_access.access_name, self.irods_access_read)
 
     def test_execute_write(self):
         """Test access setting for write/modify"""
@@ -1284,9 +1295,10 @@ class TestSetDataObjAccessTask(IRODSTaskTestBase):
             cls=SetAccessTask,
             name='Set access',
             inject={
-                'access_name': TEST_ACCESS_WRITE_IN,
+                'access_name': IRODS_ACCESS_WRITE_IN,
                 'path': self.obj_path,
                 'user_name': DEFAULT_USER_GROUP,
+                'irods_backend': self.irods_backend,
                 'obj_target': True,
             },
         )
@@ -1302,7 +1314,7 @@ class TestSetDataObjAccessTask(IRODSTaskTestBase):
             target=self._get_test_obj(), user_name=DEFAULT_USER_GROUP
         )
         self.assertIsInstance(user_access, iRODSAccess)
-        self.assertEqual(user_access.access_name, TEST_ACCESS_WRITE_OUT)
+        self.assertEqual(user_access.access_name, self.irods_access_write)
 
     def test_execute_twice(self):
         """Test access setting twice"""
@@ -1310,9 +1322,10 @@ class TestSetDataObjAccessTask(IRODSTaskTestBase):
             cls=SetAccessTask,
             name='Set access',
             inject={
-                'access_name': TEST_ACCESS_READ_IN,
+                'access_name': IRODS_ACCESS_READ_IN,
                 'path': self.obj_path,
                 'user_name': DEFAULT_USER_GROUP,
+                'irods_backend': self.irods_backend,
                 'obj_target': True,
             },
         )
@@ -1325,9 +1338,10 @@ class TestSetDataObjAccessTask(IRODSTaskTestBase):
             cls=SetAccessTask,
             name='Set access',
             inject={
-                'access_name': TEST_ACCESS_READ_IN,
+                'access_name': IRODS_ACCESS_READ_IN,
                 'path': self.obj_path,
                 'user_name': DEFAULT_USER_GROUP,
+                'irods_backend': self.irods_backend,
                 'obj_target': True,
             },
         )
@@ -1338,7 +1352,7 @@ class TestSetDataObjAccessTask(IRODSTaskTestBase):
             target=self._get_test_obj(), user_name=DEFAULT_USER_GROUP
         )
         self.assertIsInstance(user_access, iRODSAccess)
-        self.assertEqual(user_access.access_name, TEST_ACCESS_READ_OUT)
+        self.assertEqual(user_access.access_name, self.irods_access_read)
 
     def test_revert_created(self):
         """Test access setting"""
@@ -1346,9 +1360,10 @@ class TestSetDataObjAccessTask(IRODSTaskTestBase):
             cls=SetAccessTask,
             name='Set access',
             inject={
-                'access_name': TEST_ACCESS_READ_IN,
+                'access_name': IRODS_ACCESS_READ_IN,
                 'path': self.obj_path,
                 'user_name': DEFAULT_USER_GROUP,
+                'irods_backend': self.irods_backend,
                 'obj_target': True,
             },
             force_fail=True,
@@ -1367,9 +1382,10 @@ class TestSetDataObjAccessTask(IRODSTaskTestBase):
             cls=SetAccessTask,
             name='Set access',
             inject={
-                'access_name': TEST_ACCESS_READ_IN,
+                'access_name': IRODS_ACCESS_READ_IN,
                 'path': self.obj_path,
                 'user_name': DEFAULT_USER_GROUP,
+                'irods_backend': self.irods_backend,
                 'obj_target': True,
             },
         )
@@ -1380,9 +1396,10 @@ class TestSetDataObjAccessTask(IRODSTaskTestBase):
             cls=SetAccessTask,
             name='Set access',
             inject={
-                'access_name': TEST_ACCESS_WRITE_IN,
+                'access_name': IRODS_ACCESS_WRITE_IN,
                 'path': self.obj_path,
                 'user_name': DEFAULT_USER_GROUP,
+                'irods_backend': self.irods_backend,
                 'obj_target': True,
             },
             force_fail=True,
@@ -1394,7 +1411,7 @@ class TestSetDataObjAccessTask(IRODSTaskTestBase):
             target=self._get_test_obj(), user_name=DEFAULT_USER_GROUP
         )
         self.assertIsInstance(user_access, iRODSAccess)
-        self.assertEqual(user_access.access_name, TEST_ACCESS_READ_OUT)
+        self.assertEqual(user_access.access_name, self.irods_access_read)
 
     def test_revert_not_modified(self):
         """Test access setting reverting without modification"""
@@ -1402,9 +1419,10 @@ class TestSetDataObjAccessTask(IRODSTaskTestBase):
             cls=SetAccessTask,
             name='Set access',
             inject={
-                'access_name': TEST_ACCESS_READ_IN,
+                'access_name': IRODS_ACCESS_READ_IN,
                 'path': self.obj_path,
                 'user_name': DEFAULT_USER_GROUP,
+                'irods_backend': self.irods_backend,
                 'obj_target': True,
             },
         )
@@ -1416,9 +1434,10 @@ class TestSetDataObjAccessTask(IRODSTaskTestBase):
             cls=SetAccessTask,
             name='Set access',
             inject={
-                'access_name': TEST_ACCESS_READ_IN,
+                'access_name': IRODS_ACCESS_READ_IN,
                 'path': self.obj_path,
                 'user_name': DEFAULT_USER_GROUP,
+                'irods_backend': self.irods_backend,
                 'obj_target': True,
             },
             force_fail=True,
@@ -1430,7 +1449,7 @@ class TestSetDataObjAccessTask(IRODSTaskTestBase):
             target=self._get_test_obj(), user_name=DEFAULT_USER_GROUP
         )
         self.assertIsInstance(user_access, iRODSAccess)
-        self.assertEqual(user_access.access_name, TEST_ACCESS_READ_OUT)
+        self.assertEqual(user_access.access_name, self.irods_access_read)
 
 
 class TestCreateUserTask(IRODSTaskTestBase):
@@ -2021,8 +2040,9 @@ class TestBatchMoveDataObjectsTask(IRODSTaskTestBase):
                 'src_root': self.batch_src_path,
                 'dest_root': self.batch_dest_path,
                 'src_paths': [self.batch_obj_path, self.batch_obj2_path],
-                'access_name': TEST_ACCESS_READ_IN,
+                'access_name': IRODS_ACCESS_READ_IN,
                 'user_name': DEFAULT_USER_GROUP,
+                'irods_backend': self.irods_backend,
             },
         )
         self.assertFalse(self.irods.data_objects.exists(self.dest_obj_path))
@@ -2056,13 +2076,13 @@ class TestBatchMoveDataObjectsTask(IRODSTaskTestBase):
             user_name=DEFAULT_USER_GROUP,
         )
         self.assertIsInstance(obj_access, iRODSAccess)
-        self.assertEqual(obj_access.access_name, TEST_ACCESS_READ_OUT)
+        self.assertEqual(obj_access.access_name, self.irods_access_read)
         obj_access = self._get_user_access(
             target=self.irods.data_objects.get(self.dest_obj_path),
             user_name=DEFAULT_USER_GROUP,
         )
         self.assertIsInstance(obj_access, iRODSAccess)
-        self.assertEqual(obj_access.access_name, TEST_ACCESS_READ_OUT)
+        self.assertEqual(obj_access.access_name, self.irods_access_read)
 
     def test_revert(self):
         """Test reverting the moving of data objects"""
@@ -2073,8 +2093,9 @@ class TestBatchMoveDataObjectsTask(IRODSTaskTestBase):
                 'src_root': self.batch_src_path,
                 'dest_root': self.batch_dest_path,
                 'src_paths': [self.batch_obj_path, self.batch_obj2_path],
-                'access_name': TEST_ACCESS_READ_IN,
+                'access_name': IRODS_ACCESS_READ_IN,
                 'user_name': DEFAULT_USER_GROUP,
+                'irods_backend': self.irods_backend,
             },
             force_fail=True,
         )  # FAILS
@@ -2108,8 +2129,9 @@ class TestBatchMoveDataObjectsTask(IRODSTaskTestBase):
                 'src_root': self.batch_src_path,
                 'dest_root': self.batch_dest_path,
                 'src_paths': [self.batch_obj_path, self.batch_obj2_path],
-                'access_name': TEST_ACCESS_READ_IN,
+                'access_name': IRODS_ACCESS_READ_IN,
                 'user_name': DEFAULT_USER_GROUP,
+                'irods_backend': self.irods_backend,
             },
         )
         with self.assertRaises(Exception):